pgml-dashboard/requirements.txt
asgiref==3.5.0
black==22.3.0
certifi==2022.6.15
click==8.1.3
Django==4.0.7
django-extensions==3.2.0
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
sentry-sdk==1.9.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.11
whitenoise==6.0.0
PyJWT==2.4.0
Markdown==3.3.7
django-click==2.3.0
django-request==1.6.1
